Move Filter-related RemoveKey tests to FilterFixture
