+++ src/DynamicData.Tests/Cache/RemoveKeyFixture.cs
@@ -92,53 +92,4 @@ public class RemoveKeyFixture : IDisposable
         Assert.Equivalent(people, collection);
     }
 
-    [Fact]
-    public void Cache_AutoRefreshRemoveKeyFilterUpdate_CollectionUpdated()
-    {
-        var people = _generator.Take(100).ToArray();
-        var average = people.Average(x => x.Age);
-        ReadOnlyObservableCollection<Person> collection;
-        _cleanup.Add(
-            _source.Connect()
-                .AutoRefresh(x => x.Age)
-                .RemoveKey()
-                .Filter(x => x.Age < average)
-                .Bind(out collection)
-                .Subscribe()
-        );
-        _source.AddOrUpdate(people);
-
-        Assert.Equivalent(people.Where(x => x.Age < average), collection);
-
-        foreach (var person in people)
-        {
-            person.Age = person.Age + 1;
-        }
-        Assert.Equivalent(people.Where(x => x.Age < average), collection);
-    }
-
-    [Fact]
-    public void Cache_AutoRefreshFilterRemoveKeyUpdate_CollectionUpdated()
-    {
-        var people = _generator.Take(100).ToArray();
-        var average = people.Average(x => x.Age);
-        ReadOnlyObservableCollection<Person> collection;
-        _cleanup.Add(
-            _source.Connect()
-                .AutoRefresh(x => x.Age)
-                .Filter(x => x.Age < average)
-                .RemoveKey()
-                .Bind(out collection)
-                .Subscribe()
-        );
-        _source.AddOrUpdate(people);
-
-        Assert.Equivalent(people.Where(x => x.Age < average), collection);
-
-        foreach (var person in people)
-        {
-            person.Age = person.Age + 1;
-        }
-        Assert.Equivalent(people.Where(x => x.Age < average), collection);
-    }
 }
